Add new module load_lc_cores.py
